+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 
 - Enable incorrectness certificates in forward mode.
-- Enable -L/--lemmas output in forward mode.
+- Enable -l/--lemmas output in forward mode.
+- Enable -L/--lrat output in forward mode.
 
 # 0.3.0 (2020-01-01)